// src/types.d.ts
export interface ApiResponse {
    status: string;
    message?: string;
}
// ...
export interface StatusResponse extends ApiResponse {
}
export interface CoinDepositAddressNetwork {
    name: string;
    network: string;
    address: string;
    memo: string;
}
export interface CoinDepositAddressResponse extends ApiResponse {
    networks: CoinDepositAddressNetwork[];
}
export interface QuoteResponse extends ApiResponse {
    rate: string;
}
export interface PlaceOrderResponse extends ApiResponse {
    coin: string;
    market: string;
    amount: string;
    rate: string;
    id: string;
}
export interface EditOrderResponse extends ApiResponse {
    id: string;
    coin: string;
    rate: string;
    newrate: string;
    amount: string;
    total: string;
    updated: boolean;
}
export interface BuyNowOrderResponse extends ApiResponse {
    coin: string;
    amount: string;
    market: string;
    total: string;
}
export interface SellNowOrderResponse extends BuyNowOrderResponse {
    rate: string;
}
export interface SwapNowOrderResponse extends SellNowOrderResponse {
}
export interface CancelOrderResponse extends ApiResponse {
}
export interface CoinWithdrawalNetwork {
    network: string;
    paymentid: 'no' | 'optional' | 'required';
    fee: number;
    minsend: number;
    default: boolean;
}
export interface CoinWithdrawalDetailsResponse extends ApiResponse {
    networks: CoinWithdrawalNetwork[];
}